requirements.txt
ultralytics
streamlit
opencv-python
numpy
pandas
matplotlib
pillow
grad-cam
torch